backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-openai
langchain-groq
chromadb
pypdf
tiktoken